Rename scraper packages, add instagram_scraper to docker-compose file
