+++ instagram_scraper/Dockerfile
@@ -1,7 +1,7 @@
 FROM golang:1.12 as builder
 WORKDIR /go/src/github.com/codeuniversity/smag-mvp
 COPY . .
-RUN GO111MODULE=on CGO_ENABLED=0 GOOS=linux go build -installsuffix cgo -o instascraper_scraper scraper/main/main.go
+RUN GO111MODULE=on CGO_ENABLED=0 GOOS=linux go build -installsuffix cgo -o instascraper_scraper instagram_scraper/main/main.go
 
 FROM alpine
 RUN apk --no-cache add ca-certificates
