+++ twitter_scraper/Dockerfile
@@ -6,7 +6,7 @@ COPY Pipfile Pipfile.lock ./
 RUN pip install pipenv \
     && pipenv install
 
-COPY ./twitterscraper ./twitterscraper
+COPY ./twitter_scraper ./twitter_scraper
 
 ENTRYPOINT [ "pipenv", "run", "python" ]
 CMD [ "-c", "raise Eexception('Please set the CMD to either tweets_scraper.py or user_scraper.py')" ]
